+++ app/app_launcher.h
@@ -2,6 +2,5 @@
 #define APP_LAUNCHER_H
 
 void App_Launch(void);
-void App_DebugUartHandler(void);
 
 #endif /* APP_LAUNCHER_H */